Вынес обработку запросов процессора к пользователю в базовый класс
